+++ libsolidity/lsp/SemanticHighlight.cpp
@@ -26,10 +26,10 @@ using namespace solidity::frontend;
 void SemanticHighlight::operator()(MessageID _id, Json::Value const& _args)
 {
 	auto const [sourceUnitName, lineColumn] = extractSourceUnitNameAndLineColumn(_args);
-	ASTNode const* sourceNode = m_server.astNodeAtSourceLocation(sourceUnitName, lineColumn);
+	auto const [sourceNode, sourceOffset] = m_server.astNodeAndOffsetAtSourceLocation(sourceUnitName, lineColumn);
 
 	Json::Value jsonReply = Json::arrayValue;
-	for (auto const& [location, kind]: semanticHighlight(sourceNode, sourceUnitName))
+	for (auto const& [location, kind]: semanticHighlight(sourceNode, sourceOffset, sourceUnitName))
 	{
 		Json::Value item = Json::objectValue;
 		item["range"] = toRange(location);
@@ -40,13 +40,13 @@ void SemanticHighlight::operator()(MessageID _id, Json::Value const& _args)
 	client().reply(_id, jsonReply);
 }
 
-vector<Reference> SemanticHighlight::semanticHighlight(ASTNode const* _sourceNode, string const& _sourceUnitName)
+vector<Reference> SemanticHighlight::semanticHighlight(ASTNode const* _sourceNode, int _sourceOffset, string const& _sourceUnitName)
 {
 	if (!_sourceNode)
 		return {};
 
 	SourceUnit const& sourceUnit = m_server.ast(_sourceUnitName);
 
-	return ReferenceCollector::collect(_sourceNode, sourceUnit);
+	return ReferenceCollector::collect(_sourceNode, _sourceOffset, sourceUnit);
 }