+++ libsolidity/lsp/SemanticHighlight.h
@@ -23,15 +23,18 @@
 namespace solidity::lsp
 {
 
+/**
+ * Implements JSON RPC for `textDocument/documentHighlight`.
+ */
 class SemanticHighlight: public HandlerBase
 {
 public:
-	SemanticHighlight(LanguageServer& _server): HandlerBase(_server) {}
+	explicit SemanticHighlight(LanguageServer& _server): HandlerBase(_server) {}
 
 	void operator()(MessageID _id, Json::Value const& _args);
 
 private:
-	std::vector<Reference> semanticHighlight(frontend::ASTNode const* _sourceNode, std::string const& _sourceUnitName);
+	std::vector<Reference> semanticHighlight(frontend::ASTNode const* _sourceNode, int _sourceOffset, std::string const& _sourceUnitName);
 };
 
 }